Added some trees with some different colors
Also changed the speeds for the camera to make it fly a bit faster.
changed the name of the window to "fly, you fools" haha.  All the tree
fractal calculations are done once and put in a list to keep everything
fast.  Framework is in place to make leaves different colors.